# docs/swagger.yaml
definitions:
  models.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        enum:
        - ADMIN
        - USER
        type: string
    required:
    - email
    - password
    - role
    type: object
  models.LoginResponse:
    properties:
      token:
        type: string
      userId:
        type: string
    type: object
  models.RegisterInput:
    properties:
      confirmPassword:
        type: string
      email:
        type: string
      password:
        type: string
      token:
        type: string
    required:
    - confirmPassword
    - email
    - password
    - token
    type: object
  models.RegisterResponse:
    properties:
      message:
        type: string
    type: object
  models.SendEmailOtpInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  models.SendEmailResponse:
    properties:
      otp:
        description: 'todo: remove otp'
        type: string
      token:
        type: string
    type: object
  models.VerifyOtpInput:
    properties:
      email:
        type: string
      otp:
        type: string
      token:
        type: string
    required:
    - email
    - otp
    - token
    type: object
  models.VerifyOtpResponse:
    properties:
      token:
        type: string
    type: object
  orders.OrderCancelResponse:
    properties:
      message:
        type: string
    type: object
  orders.OrderCreateInput:
    properties:
      orderId:
        type: integer
      productId:
        type: integer
      quantity:
        type: integer
    type: object
  orders.OrderCreateResponse:
    properties:
      message:
        type: string
    type: object
  orders.OrderGetAllResponse:
    properties:
      id:
        type: string
      product1:
        type: string
      product2:
        type: string
      product3:
        type: string
    type: object
  orders.OrderGetOneResponse:
    properties:
      id:
        type: integer
      productId:
        type: integer
      quantity:
        type: integer
    type: object
  orders.OrderUpdateStatusInput:
    properties:
      status:
        enum:
        - CANCEL
        type: string
    type: object
  products.ProductCreateInput:
    properties:
      description:
        type: string
      imageUrl:
        type: string
      price:
        type: integer
      stock:
        type: integer
      title:
        type: string
    required:
    - description
    - price
    - stock
    - title
    type: object
  products.ProductCreateResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      imageUrl:
        type: string
      price:
        type: integer
      title:
        type: string
    type: object
  products.ProductDeleteResponse:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
  products.ProductGetAllResponse:
    properties:
      description:
        type: string
      id:
        type: string
      price:
        type: string
      title:
        type: string
    type: object
  products.ProductGetOneResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      price:
        type: integer
      title:
        type: string
    type: object
  products.ProductUpdateInput:
    properties:
      description:
        type: string
      imageUrl:
        type: string
      price:
        type: integer
      stock:
        type: integer
      title:
        type: string
    required:
    - description
    - price
    - stock
    - title
    type: object
  products.ProductUpdateResponse:
    properties:
      message:
        type: string
    type: object
  types.ErrMsg:
    properties:
      error:
        type: string
    type: object
host: localhost:1323
info:
  contact:
    email: instashop@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a simple e-commerce server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Instashop Swagger API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: '"You can either login as a USER or an ADMIN. Case-sensitive roles"'
      parameters:
      - description: Auth_Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: error
          schema:
            $ref: '#/definitions/types.ErrMsg'
      tags:
      - Auth_Login
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Auth_Register
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: success
          schema:
            $ref: '#/definitions/models.RegisterResponse'
        "400":
          description: error
          schema:
            $ref: '#/definitions/types.ErrMsg'
      tags:
      - Auth_Register
  /auth/send-otp:
    post:
      consumes:
      - application/json
      parameters:
      - description: Auth_SendEmailOtp
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SendEmailOtpInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/models.SendEmailResponse'
        "400":
          description: error
          schema:
            $ref: '#/definitions/types.ErrMsg'
      tags:
      - Auth_SendEmailOtp
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      parameters:
      - description: Auth_VerifyOtp
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.VerifyOtpInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/models.VerifyOtpResponse'
        "400":
          description: error
          schema:
            $ref: '#/definitions/types.ErrMsg'
      tags:
      - Auth_VerifyOtp
  /orders:
    get:
      consumes:
      - application/json
      parameters:
      - description: Header must be set for valid response
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            items:
              $ref: '#/definitions/orders.OrderGetAllResponse'
            type: array
      tags:
      - Order_GetAll
    post:
      consumes:
      - application/json
      parameters:
      - description: Header must be set for valid response
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order_Create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/orders.OrderCreateInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/orders.OrderCreateResponse'
      tags:
      - Order_Create
  /orders/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id of order
        in: path
        name: id
        required: true
        type: string
      - description: Header must be set for valid response
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - Order_Delete
    get:
      consumes:
      - application/json
      parameters:
      - description: id of order
        in: path
        name: id
        required: true
        type: string
      - description: Header must be set for valid response
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            items:
              $ref: '#/definitions/orders.OrderGetOneResponse'
            type: array
      tags:
      - Order_GetOne
    put:
      consumes:
      - application/json
      parameters:
      - description: id of order
        in: path
        name: id
        required: true
        type: string
      - description: Header must be set for valid response
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - Order_Update
  /orders/cancel/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: id of order
        in: path
        name: id
        required: true
        type: string
      - description: Order_UpdateStatus
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/orders.OrderUpdateStatusInput'
      - description: Header must be set for valid response
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/orders.OrderCancelResponse'
        "400":
          description: error
          schema:
            $ref: '#/definitions/types.ErrMsg'
      tags:
      - Order_UpdateStatus
  /products:
    get:
      consumes:
      - application/json
      parameters:
      - description: Header must be set for valid response
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            items:
              $ref: '#/definitions/products.ProductGetAllResponse'
            type: array
      tags:
      - Product_GetAll
    post:
      consumes:
      - application/json
      parameters:
      - description: Product_Create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/products.ProductCreateInput'
      - description: Header must be set for valid response
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/products.ProductCreateResponse'
      tags:
      - Product_Create
  /products/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id of product
        in: path
        name: id
        required: true
        type: string
      - description: Header must be set for valid response
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/products.ProductDeleteResponse'
      tags:
      - Product_Delete
    get:
      consumes:
      - application/json
      parameters:
      - description: id of product
        in: path
        name: id
        required: true
        type: string
      - description: Header must be set for valid response
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/products.ProductGetOneResponse'
      tags:
      - Product_GetOne
    patch:
      consumes:
      - application/json
      parameters:
      - description: id of product
        in: path
        name: id
        required: true
        type: string
      - description: Header must be set for valid response
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product_Update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/products.ProductUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/products.ProductUpdateResponse'
      tags:
      - Product_Update
swagger: "2.0"